+++ plugins.ts
@@ -2,9 +2,9 @@ import postcss from "lume/plugins/postcss.ts";
 import pagefind from "lume/plugins/pagefind.ts";
 import resolveUrls from "lume/plugins/resolve_urls.ts";
 import nav from "lume/plugins/nav.ts";
-import title from "https://deno.land/x/lume_markdown_plugins@v0.7.0/title.ts";
-import toc from "https://deno.land/x/lume_markdown_plugins@v0.7.0/toc.ts";
-import footnotes from "https://deno.land/x/lume_markdown_plugins@v0.7.0/footnotes.ts";
+import title from "https://deno.land/x/lume_markdown_plugins@v0.7.1/title.ts";
+import toc from "https://deno.land/x/lume_markdown_plugins@v0.7.1/toc.ts";
+import footnotes from "https://deno.land/x/lume_markdown_plugins@v0.7.1/footnotes.ts";
 import prism from "lume/plugins/prism.ts";
 import date from "lume/plugins/date.ts";
 import basePath from "lume/plugins/base_path.ts";
@@ -12,7 +12,7 @@ import favicon, { Options as FaviconOptions } from "lume/plugins/favicon.ts";
 import phosphor, {
   Options as IconOptions,
 } from "https://deno.land/x/lume_icon_plugins@v0.2.3/phosphor.ts";
-import { alert } from "npm:@mdit/plugin-alert@0.8.0";
+import { alert } from "npm:@mdit/plugin-alert@0.13.1";
 
 import "lume/types.ts";