+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.11.0] - Unreleased
+## [0.11.0] - 2024-09-10
 ### Added
+- Support for Lume 2.3.2
 - Support for multilanguage sites, with the option `languages` and `languageNames`.
 - New variable `hide_menu` to hide a page from the main menu.
 
@@ -162,7 +163,7 @@ First version
 [#8]: https://github.com/lumeland/theme-simple-wiki/issues/8
 [#9]: https://github.com/lumeland/theme-simple-wiki/issues/9
 
-[0.11.0]: https://github.com/lumeland/theme-simple-wiki/compare/v0.10.1...HEAD
+[0.11.0]: https://github.com/lumeland/theme-simple-wiki/compare/v0.10.1...v0.11.0
 [0.10.1]: https://github.com/lumeland/theme-simple-wiki/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/lumeland/theme-simple-wiki/compare/v0.9.3...v0.10.0
 [0.9.3]: https://github.com/lumeland/theme-simple-wiki/compare/v0.9.2...v0.9.3